+++ src/transport/mqtt/pairing.rs
@@ -99,45 +99,28 @@ pub(crate) struct ApiClient<'a> {
     pub(crate) device_id: &'a str,
     pairing_url: &'a Url,
     credentials_secret: &'a str,
+    client: reqwest::Client,
 }
 
 impl<'a> ApiClient<'a> {
-    fn new(
-        realm: &'a str,
-        device_id: &'a str,
-        pairing_url: &'a Url,
-        credentials_secret: &'a str,
-    ) -> Self {
-        // TODO: remove this when we can configure TlsClient
-        #[cfg(test)]
-        if rustls::crypto::CryptoProvider::get_default().is_none() {
-            tracing::debug!("no provider set, installing default provider");
-
-            rustls::crypto::CryptoProvider::install_default(
-                rustls::crypto::aws_lc_rs::default_provider(),
-            )
-            .expect("failed to install default provider");
-        }
-
-        Self {
-            realm,
-            device_id,
-            pairing_url,
-            credentials_secret,
-        }
-    }
-
     pub(crate) fn from_transport(
         provider: &'a TransportProvider,
         realm: &'a str,
         device_id: &'a str,
-    ) -> Self {
-        Self::new(
+    ) -> Result<Self, PairingError> {
+        let tls_config = provider.api_tls_config()?;
+
+        let client = reqwest::Client::builder()
+            .use_preconfigured_tls(tls_config.clone())
+            .build()?;
+
+        Ok(Self {
             realm,
             device_id,
-            provider.pairing_url(),
-            provider.credential_secret(),
-        )
+            pairing_url: provider.pairing_url(),
+            credentials_secret: provider.credential_secret(),
+            client,
+        })
     }
 
     fn url<'i, I>(&self, segments: I) -> Result<Url, PairingError>
@@ -169,8 +152,8 @@ impl<'a> ApiClient<'a> {
 
         let payload = ApiData::new(MqttV1Csr { csr });
 
-        let client = reqwest::Client::new();
-        let response = client
+        let response = self
+            .client
             .post(url)
             .bearer_auth(self.credentials_secret)
             .json(&payload)
@@ -197,8 +180,8 @@ impl<'a> ApiClient<'a> {
     pub async fn get_broker_url(&self) -> Result<Url, PairingError> {
         let url = self.url([])?;
 
-        let client = reqwest::Client::new();
-        let response = client
+        let response = self
+            .client
             .get(url)
             .bearer_auth(self.credentials_secret)
             .send()
@@ -342,7 +325,13 @@ pub(crate) mod tests {
         let mock = mock_create_certificate(&mut server).create_async().await;
 
         let parse = Url::parse(&server.url()).unwrap();
-        let client = ApiClient::new("realm", "device_id", &parse, "secret");
+
+        let provider = TransportProvider::configure(parse, "secret".to_string(), None, true)
+            .await
+            .expect("couldn't configure provider");
+
+        let client = ApiClient::from_transport(&provider, "realm", "device_id")
+            .expect("couldn't create api client");
 
         let bundle = Bundle::new("test", "device_id").unwrap();
 
@@ -373,7 +362,13 @@ pub(crate) mod tests {
             .await;
 
         let url = Url::parse(&server.url()).unwrap();
-        let client = ApiClient::new("realm", "device_id", &url, "secret");
+
+        let provider = TransportProvider::configure(url, "secret".to_string(), None, true)
+            .await
+            .expect("couldn't configure provider");
+
+        let client = ApiClient::from_transport(&provider, "realm", "device_id")
+            .expect("couldn't create api client");
 
         let res = client
             .create_certificate("csr")
@@ -392,7 +387,13 @@ pub(crate) mod tests {
         let mock = mock_get_broker_url(&mut server).create_async().await;
 
         let url = Url::parse(&server.url()).unwrap();
-        let client = ApiClient::new("realm", "device_id", &url, "secret");
+
+        let provider = TransportProvider::configure(url, "secret".to_string(), None, true)
+            .await
+            .expect("couldn't configure provider");
+
+        let client = ApiClient::from_transport(&provider, "realm", "device_id")
+            .expect("couldn't create api client");
 
         let res = client.get_broker_url().await.unwrap();
 
@@ -415,7 +416,13 @@ pub(crate) mod tests {
             .await;
 
         let url = Url::parse(&server.url()).unwrap();
-        let client = ApiClient::new("realm", "device_id", &url, "secret");
+
+        let provider = TransportProvider::configure(url, "secret".to_string(), None, true)
+            .await
+            .expect("couldn't configure provider");
+
+        let client = ApiClient::from_transport(&provider, "realm", "device_id")
+            .expect("couldn't create api client");
 
         let res = client.get_broker_url().await.expect_err("should error");