+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+### Changed
+
+- Use a default CryptoProvider when defining a Tls configuration
+  [#425](https://github.com/astarte-platform/astarte-device-sdk-rust/pull/425).
+
 ## [0.9.5] - 2025-03-04
 
 ### Changed